+++ src/editor/core/menu/menuitem.js
@@ -105,9 +105,6 @@ export class MenuItem {
             e.preventDefault();
             if (!this.$.hasClass(buildMenuClass('disabled'))) {
                 this.options.run.call(this, view.state, view.dispatch, view, e);
-                if (!this.$.is(':visible')) {
-                    this.getMenuBar().focusPrev();
-                }
             }
         };