+++ src/editor/core/plugins/code_block/keymap.js
@@ -3,12 +3,9 @@ import {Selection} from 'prosemirror-state';
 function arrowHandler(dir) {
     return (state, dispatch, view) => {
         if (state.selection.empty && view.endOfTextblock(dir)) {
-            let side = dir === "left" || dir === "up" ? -1 : 1,
-                $head = state.selection.$head;
-            let nextPos = Selection.near(
-                state.doc.resolve(side > 0 ? $head.after() : $head.before()),
-                side
-            );
+            const side = dir === "left" || dir === "up" ? -1 : 1, $head = state.selection.$head;
+            const pos = side > 0 ? $head.after() : $head.before();
+            const nextPos = Selection.near(state.doc.resolve(pos), side);
             if (nextPos.$head && nextPos.$head.parent.type.name === "code_block") {
                 dispatch(state.tr.setSelection(nextPos));
                 return true;
@@ -18,7 +15,7 @@ function arrowHandler(dir) {
     };
 }
 
-let keymap = () => {
+const keymap = () => {
     return {
         ArrowLeft: arrowHandler("left"),
         ArrowRight: arrowHandler("right"),