+++ src/editor/core/plugins/em/schema.js
@@ -2,7 +2,6 @@
  * @link https://www.humhub.org/
  * @copyright Copyright (c) 2017 HumHub GmbH & Co. KG
  * @license https://www.humhub.com/licences
- *
  */
 
 const schema = {
@@ -12,7 +11,7 @@ const schema = {
             parseDOM: [{tag: "i"}, {tag: "em"},
                 {
                     style: "font-style", getAttrs: function (value) {
-                    return value == "italic" && null;
+                    return value === "italic" && null;
                 }
                 }],
             toDOM: () => {
@@ -24,4 +23,4 @@ const schema = {
     }
 };
 
-export {schema}
+export {schema};