+++ test/menu/test-menu-accessibility.js
@@ -173,7 +173,7 @@ describe("Menu:accessibility", () => {
         // navigate(DIR_DOWN, 'makeHeading6');
         navigate(DIR_DOWN, 'makeCodeBlock');
         pressEnter();
-        expect(toHtml()).to.equal('<div class="cm-editor cm-focused ͼ1 ͼ2 ͼ4" contenteditable="false"><div aria-live="polite" style="position: fixed; top: -10000px;"></div><div tabindex="-1" class="cm-scroller"><div class="cm-gutters" aria-hidden="true" style="min-height: 14px; position: sticky;"><div class="cm-gutter cm-lineNumbers"><div class="cm-gutterElement" style="height: 0px; visibility: hidden; pointer-events: none;">9</div><div class="cm-gutterElement" style="height: 14px;">1</div></div></div><div spellcheck="false" autocorrect="off" autocapitalize="off" translate="no" contenteditable="true" class="cm-content" style="tab-size: 4" role="textbox" aria-multiline="true" data-language="javascript"><div class="cm-line"><br></div></div><div class="cm-layer cm-layer-above cm-cursorLayer" aria-hidden="true" style="z-index: 150; animation-duration: 1200ms; animation-name: cm-blink;"></div><div class="cm-layer cm-selectionLayer" aria-hidden="true" style="z-index: -2;"></div></div></div>');
+        expect(toHtml('#stage .ProseMirror .cm-editor .cm-content')).to.equal('<div class="cm-line"><br></div>');
         done();
     });